# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CCN War Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CcnEquipo {
  id: string;
  name: string;
  /** @nullable */
  clanTag?: string | null;
  /** @nullable */
  clanName?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  elo?: number | null;
  /** @nullable */
  rank?: number | null;
  /** @nullable */
  wins?: number | null;
  /** @nullable */
  losses?: number | null;
  /** @nullable */
  badgeUrl?: string | null;
}

export interface CcnPartida {
  id: string;
  homeTeam: CcnEquipo;
  awayTeam: CcnEquipo;
  /** @nullable */
  homeStars?: number | null;
  /** @nullable */
  awayStars?: number | null;
  /** @nullable */
  homeDestruction?: number | null;
  /** @nullable */
  awayDestruction?: number | null;
  /** @nullable */
  scheduledAt?: string | null;
  status: string;
  /** @nullable */
  tournamentName?: string | null;
  /** @nullable */
  winner?: string | null;
}

export type CcnPartidaList = CcnPartida[];

export interface CcnGuerraItem {
  id: string;
  homeTeam: string;
  awayTeam: string;
  /** @nullable */
  homeTag?: string | null;
  /** @nullable */
  awayTag?: string | null;
  scheduledAt: string;
  scheduledAtSv: string;
  scheduledAtAr: string;
  status: string;
  /** @nullable */
  tournamentName?: string | null;
  /** @nullable */
  homeStars?: number | null;
  /** @nullable */
  awayStars?: number | null;
  /** @nullable */
  homeDestruction?: number | null;
  /** @nullable */
  awayDestruction?: number | null;
  /** @nullable */
  winner?: string | null;
}

export type CcnGuerraList = CcnGuerraItem[];

export interface CcnRankingEloItem {
  rank: number;
  teamId: string;
  teamName: string;
  elo: number;
  wins: number;
  losses: number;
  /** @nullable */
  clanTag?: string | null;
  /** @nullable */
  badgeUrl?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  profileUrl?: string | null;
  /** @nullable */
  eloClassic?: number | null;
  /** @nullable */
  rankClassic?: number | null;
}

export type CcnRankingEloList = CcnRankingEloItem[];

export interface CcnRankingMensualItem {
  rank: number;
  teamId: string;
  teamName: string;
  points: number;
  wins: number;
  losses: number;
  /** @nullable */
  clanTag?: string | null;
}

export type CcnRankingMensualList = CcnRankingMensualItem[];

export interface CcnTorneo {
  id: string;
  name: string;
  status: string;
  /** @nullable */
  startDate?: string | null;
  /** @nullable */
  endDate?: string | null;
  /** @nullable */
  region?: string | null;
  /** @nullable */
  teamsCount?: number | null;
}

export type CcnTorneoList = CcnTorneo[];

export type GetCcnGuerrasParams = {
/**
 * 0 = today, 1 = tomorrow
 */
offset?: number;
};

export type SearchCcnEquipoParams = {
q: string;
};