# lib/api-zod/src/generated/types/ccnRankingEloItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CCN War Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CcnRankingEloItem {
  rank: number;
  teamId: string;
  teamName: string;
  elo: number;
  wins: number;
  losses: number;
  /** @nullable */
  clanTag?: string | null;
  /** @nullable */
  badgeUrl?: string | null;
  /** @nullable */
  twitter?: string | null;
  /** @nullable */
  profileUrl?: string | null;
  /** @nullable */
  eloClassic?: number | null;
  /** @nullable */
  rankClassic?: number | null;
}